+++ CHANGELOG.md
@@ -4,8 +4,22 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.4.0-pre.3] - 2024-04-09
+
+### Fixed
+- Outdated documentation.
+- Error message for importing incompatible Sentis models.
+- NonMaxSuppression output tensor has correct shape.
+- Output of Size operator.
+
+### Removed
+- Non-working NonMaxSuppression GPUCompute optimizations.
+
 ## [1.4.0-pre.2] - 2024-04-06
-- Test fixes
+
+### Fixed
+- Broken tests.
+- Outdated documentation.
 
 ## [1.4.0-pre.1] - 2024-04-05